pyproject.toml
[tool.poetry]
name = "skyhealth"
version = "0.1.0"
description = "Weather ETL platform"
authors = ["Skyhealth <engineering@skyhealth.local>"]
packages = [{include = "spark"}, {include = "apps"}, {include = "skyhealth"}]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
dagster = "^1.9.3"
dagster-graphql = "^1.9.3"
dagster-webserver = "^1.9.3"
pyspark = "^4.0.0"
delta-spark = "^4.0.0"
streamlit = "^1.49.0"
folium = "^0.20.0"
google-cloud-storage = "^2.19.0"
google-cloud-bigquery = "^3.36.0"
google-cloud-compute = "^1.37.0"
pydantic = "^2.9.2"
duckdb = "^1.3.2"
pandas = "^2.2.3"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^24.10.0"
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
